fix: improve error handling and validation in collection functions an… (#19)
* fix: improve error handling and validation in collection functions and image processing

* fix: ensure proper closure of OffscreenCanvas to prevent memory leaks

+++ PluckIt.Infrastructure/WardrobeRepository.cs
@@ -78,13 +78,13 @@ public class WardrobeRepository : IWardrobeRepository
     if (query.PriceMin.HasValue)
     {
       conditions.Add("c.price.amount >= @priceMin");
-      parameters.Add("@priceMin", (double)query.PriceMin.Value);
+      parameters.Add("@priceMin", query.PriceMin.Value);
     }
 
     if (query.PriceMax.HasValue)
     {
       conditions.Add("c.price.amount <= @priceMax");
-      parameters.Add("@priceMax", (double)query.PriceMax.Value);
+      parameters.Add("@priceMax", query.PriceMax.Value);
     }
 
     if (query.MinWears.HasValue)
